fix: rename timeline repositories path
