+++ packages/twenty-server/src/engine/object-metadata-repository/metadata-to-repository.mapping.ts
@@ -1,5 +1,5 @@
 import { BlocklistRepository } from 'src/modules/blocklist/repositories/blocklist.repository';
-import { TimelineActivityRepository } from 'src/modules/timeline/repositiories/timeline-activity.repository';
+import { TimelineActivityRepository } from 'src/modules/timeline/repositories/timeline-activity.repository';
 
 export const metadataToRepositoryMapping = {
   BlocklistWorkspaceEntity: BlocklistRepository,
